+++ CHANGELOG.md
@@ -203,3 +203,10 @@
 * EPMRPP-31435 - 'Ignore in AA' should not to be set for test item in launch on DEBUG level
 * EPMRPP-31447 - WS: "ActionType"=analyze_item for action that was done by user
 * Fix issue with incorrect importing of some junit files
+
+## 4.x.x
+##### Released: XXX XX, 2018
+
+### New Features
+
+* EPMRPP-31780 - Set the name of analyzer which made the changes